refactoring of column type specific parsing into column types
