+++ core/src/main/java/tech/tablesaw/columns/booleans/BooleanColumnType.java
@@ -1,15 +1,14 @@
 package tech.tablesaw.columns.booleans;
 
-import com.google.common.collect.Lists;
 import tech.tablesaw.api.BooleanColumn;
 import tech.tablesaw.api.ColumnType;
 import tech.tablesaw.columns.AbstractColumnType;
 import tech.tablesaw.columns.Column;
-import tech.tablesaw.columns.StringParser;
-import tech.tablesaw.io.TypeUtils;
 import tech.tablesaw.io.csv.CsvReadOptions;
 
 public class BooleanColumnType extends AbstractColumnType {
+
+    public static BooleanStringParser DEFAULT_PARSER = new BooleanStringParser(ColumnType.BOOLEAN);
 
     public BooleanColumnType(Comparable<?> missingValue, int byteSize, String name, String printerFriendlyName) {
         super(missingValue, byteSize, name, printerFriendlyName);
@@ -22,40 +21,11 @@ public class BooleanColumnType extends AbstractColumnType {
 
     @Override
     public BooleanStringParser defaultParser() {
-        return new BooleanStringParser(this);
+        return DEFAULT_PARSER;
     }
 
     @Override
     public BooleanStringParser customParser(CsvReadOptions readOptions) {
         return new BooleanStringParser(this, readOptions);
     }
-
-    static class BooleanStringParser extends StringParser<Boolean> {
-
-        public BooleanStringParser(ColumnType columnType) {
-            super(columnType);
-        }
-
-        public BooleanStringParser(BooleanColumnType booleanColumnType, CsvReadOptions readOptions) {
-            super(booleanColumnType);
-            if (readOptions.missingValueIndicator() != null) {
-                missingValueStrings = Lists.newArrayList(readOptions.missingValueIndicator());
-            }
-        }
-
-        @Override
-        public boolean canParse(String s) {
-            if (isMissing(s)) {
-                return true;
-            }
-            return TypeUtils.TRUE_STRINGS_FOR_DETECTION.contains(s)
-                    || TypeUtils.FALSE_STRINGS_FOR_DETECTION.contains(s);
-        }
-
-        @Override
-        public Boolean parse(String s) {
-            return null;
-        }
-    }
-
 }
